+++ tools/tests/test_bundler_npm.js
@@ -172,12 +172,12 @@ assert.doesNotThrow(function () {
 
   // while bundling, verify that we don't call `npm install
   // name@version unnecessarily` -- calling `npm install` is enough,
-  // and installing each package separately coul unintentionally bump
+  // and installing each package separately could unintentionally bump
   // subdependency versions. (to intentionally bump subdependencies,
   // just remove all of the .npm directory)
   var bareExecFileSync = meteorNpm._execFileSync;
   meteorNpm._execFileSync = function(file, args, opts) {
-    if (args[0] === 'install' && args[1])
+    if (args.length > 2 && args[0] === 'install' && args[1] === '--force')
       assert.fail("shouldn't be installing specific npm packages: " + args[1]);
     return bareExecFileSync(file, args, opts);
   };